Refs #16986 - Fix comments
