+++ webpack/__tests__/helpers.test.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { testSelectorsSnapshotWithFixtures } from 'react-redux-test-utils';
 import { deepPropsToCamelCase, withProtectedView } from '../helpers';
+import withProtectedView from '../withProtectedView';
 
 const ProtectedComponent = () => <div>Protected component</div>;
