+++ webpack/components/TemplateSyncResult/TemplateSyncResultHelpers.js
@@ -1,0 +1,6 @@
+// move to some sort of pagination helper in core
+export const templatesPage = (templates, pagination) => {
+  const offset = (pagination.page - 1) * pagination.perPage;
+
+  return templates.slice(offset, offset + pagination.perPage);
+};
