+++ webpack/components/TemplateSyncResult/TemplateSyncResultSelectors.js
@@ -1,8 +1,1 @@
-// move to some sort of pagination helper in core
-export const templatesPage = (templates, pagination) => {
-  const offset = (pagination.page - 1) * pagination.perPage;
-
-  return templates.slice(offset, offset + pagination.perPage);
-};
-
 export const selectSyncResult = state => state.foremanTemplates.syncResult;
