+++ webpack/components/TemplateSyncResult/components/SyncedTemplate/InfoItem.js
